requirements.txt
fastapi
pyyaml
uvicorn
ollama
openai